memory issues fixed, caused by buffers not being deleted. Data size reduction also done.

+++ VoxelWorld.cpp
@@ -8,10 +8,10 @@ VoxelWorld::VoxelWorld() {
 
 }
 
-void VoxelWorld::LoadChunk(vec3 chunkPos){
+void VoxelWorld::LoadChunk(ivec3 chunkPos){
 
     // Find the chunk data
-    std::tuple<float, float, float> posTuple(chunkPos.x, chunkPos.y, chunkPos.z);
+    std::tuple<int, int, int> posTuple(chunkPos.x, chunkPos.y, chunkPos.z);
     Chunk* chunk = chunksMap[posTuple];
 
     // Update the new chunks neighbouring chunks and the neighbouring chunks' data about this new chunk
@@ -24,10 +24,10 @@ void VoxelWorld::LoadChunk(vec3 chunkPos){
     loadedChunks[posTuple] = true;
 }
 
-void VoxelWorld::UnloadChunk(vec3 chunkPos){
+void VoxelWorld::UnloadChunk(ivec3 chunkPos){
 
     // Find the chunk data
-    std::tuple<float, float, float> posTuple(chunkPos.x, chunkPos.y, chunkPos.z);
+    std::tuple<int, int, int> posTuple(chunkPos.x, chunkPos.y, chunkPos.z);
     Chunk* chunk = chunksMap[posTuple];
 
     // Set the loadedChunks value to false
@@ -37,7 +37,7 @@ void VoxelWorld::UnloadChunk(vec3 chunkPos){
 void VoxelWorld::AddChunk(Chunk* chunk) {
 
     vec3 chunkPos = chunk->GetPosition();
-    std::tuple<float, float, float> posTuple(chunkPos.x, chunkPos.y, chunkPos.z);
+    std::tuple<int, int, int> posTuple(chunkPos.x, chunkPos.y, chunkPos.z);
 
     chunksMap[posTuple] = chunk;
 }
@@ -53,12 +53,12 @@ void VoxelWorld::SetGreedyMeshingComputeProgramID(GLuint greedyMeshComputeProgra
 }
 
 // Returns an array neighbouring chunks load status in order - (front, right, back, left, bottom, top)
-std::vector<Chunk*> VoxelWorld::CheckNeighbouringChunksLoaded(vec3 chunkPos) {
+std::vector<Chunk*> VoxelWorld::CheckNeighbouringChunksLoaded(ivec3 chunkPos) {
 
     std::vector<Chunk*> tempChunkLoaded;
 
     // Front
-    std::tuple<float, float, float> frontChunkPosTuple(chunkPos.x, chunkPos.y, chunkPos.z + 16); // 16 should be chunk size
+    std::tuple<int, int, int> frontChunkPosTuple(chunkPos.x, chunkPos.y, chunkPos.z + 16); // 16 should be chunk size
     if(loadedChunks.find(frontChunkPosTuple) != loadedChunks.end()){
         loadedChunks[frontChunkPosTuple] ? tempChunkLoaded.push_back(chunksMap[frontChunkPosTuple]) : tempChunkLoaded.push_back(nullptr);
     }else{
@@ -66,7 +66,7 @@ std::vector<Chunk*> VoxelWorld::CheckNeighbouringChunksLoaded(vec3 chunkPos) {
     }
 
     // Right
-    std::tuple<float, float, float> rightChunkPosTuple(chunkPos.x + 16, chunkPos.y, chunkPos.z);
+    std::tuple<int, int, int> rightChunkPosTuple(chunkPos.x + 16, chunkPos.y, chunkPos.z);
     if(loadedChunks.find(rightChunkPosTuple) != loadedChunks.end()){
         loadedChunks[rightChunkPosTuple] ? tempChunkLoaded.push_back(chunksMap[rightChunkPosTuple]) : tempChunkLoaded.push_back(nullptr);
     }else{
@@ -74,7 +74,7 @@ std::vector<Chunk*> VoxelWorld::CheckNeighbouringChunksLoaded(vec3 chunkPos) {
     }
 
     // Back
-    std::tuple<float, float, float> backChunkPosTuple(chunkPos.x, chunkPos.y, chunkPos.z - 16);
+    std::tuple<int, int, int> backChunkPosTuple(chunkPos.x, chunkPos.y, chunkPos.z - 16);
     if(loadedChunks.find(backChunkPosTuple) != loadedChunks.end()){
         loadedChunks[backChunkPosTuple] ? tempChunkLoaded.push_back(chunksMap[backChunkPosTuple]) : tempChunkLoaded.push_back(nullptr);
     }else{
@@ -82,7 +82,7 @@ std::vector<Chunk*> VoxelWorld::CheckNeighbouringChunksLoaded(vec3 chunkPos) {
     }
 
     // Left
-    std::tuple<float, float, float> leftChunkPosTuple(chunkPos.x - 16, chunkPos.y, chunkPos.z);
+    std::tuple<int, int, int> leftChunkPosTuple(chunkPos.x - 16, chunkPos.y, chunkPos.z);
     if(loadedChunks.find(leftChunkPosTuple) != loadedChunks.end()){
         loadedChunks[leftChunkPosTuple] ? tempChunkLoaded.push_back(chunksMap[leftChunkPosTuple]) : tempChunkLoaded.push_back(nullptr);
     }else{
@@ -90,7 +90,7 @@ std::vector<Chunk*> VoxelWorld::CheckNeighbouringChunksLoaded(vec3 chunkPos) {
     }
 
     // Bottom
-    std::tuple<float, float, float> botChunkPosTuple(chunkPos.x, chunkPos.y - 16, chunkPos.z);
+    std::tuple<int, int, int> botChunkPosTuple(chunkPos.x, chunkPos.y - 16, chunkPos.z);
     if(loadedChunks.find(botChunkPosTuple) != loadedChunks.end()){
         loadedChunks[botChunkPosTuple] ? tempChunkLoaded.push_back(chunksMap[botChunkPosTuple]) : tempChunkLoaded.push_back(nullptr);
     }else{
@@ -98,7 +98,7 @@ std::vector<Chunk*> VoxelWorld::CheckNeighbouringChunksLoaded(vec3 chunkPos) {
     }
 
     // Top
-    std::tuple<float, float, float> topChunkPosTuple(chunkPos.x, chunkPos.y + 16, chunkPos.z);
+    std::tuple<int, int, int> topChunkPosTuple(chunkPos.x, chunkPos.y + 16, chunkPos.z);
     if(loadedChunks.find(topChunkPosTuple) != loadedChunks.end()){
         loadedChunks[topChunkPosTuple] ? tempChunkLoaded.push_back(chunksMap[topChunkPosTuple]) : tempChunkLoaded.push_back(nullptr);
     }else{
@@ -108,13 +108,13 @@ std::vector<Chunk*> VoxelWorld::CheckNeighbouringChunksLoaded(vec3 chunkPos) {
     return tempChunkLoaded;
 }
 
-std::map<std::tuple<float, float, float>, Chunk*> VoxelWorld::GetChunksMap() {
+std::map<std::tuple<int, int, int>, Chunk*> VoxelWorld::GetChunksMap() {
 
     return chunksMap;
 }
 
-Chunk* VoxelWorld::GetChunk(vec3 chunkPos) {
+Chunk* VoxelWorld::GetChunk(ivec3 chunkPos) {
 
-    std::tuple<float, float, float> chunkPosTuple(chunkPos.x, chunkPos.y, chunkPos.z);
+    std::tuple<int, int, int> chunkPosTuple(chunkPos.x, chunkPos.y, chunkPos.z);
     return chunksMap[chunkPosTuple];
 }
